refactor(units): SI mass header (#4)
