+++ posu/units/si/ipp/si.ipp
@@ -3,187 +3,6 @@ namespace posu::units {
 
     inline namespace literals {
 
-        inline namespace mass_literals {
-
-            [[nodiscard]] constexpr auto operator""_ag(unsigned long long value) -> attograms
-            {
-                return attograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_ag(long double value) -> mass<double, std::atto>
-            {
-                return mass<double, std::atto>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_fg(unsigned long long value) -> femtograms
-            {
-                return femtograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_fg(long double value)
-                -> mass<double, std::femto>
-            {
-                return mass<double, std::femto>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_pg(unsigned long long value) -> picograms
-            {
-                return picograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_pg(long double value) -> mass<double, std::pico>
-            {
-                return mass<double, std::pico>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_ng(unsigned long long value) -> nanograms
-            {
-                return nanograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_ng(long double value) -> mass<double, std::nano>
-            {
-                return mass<double, std::nano>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_ug(unsigned long long value) -> micrograms
-            {
-                return micrograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_ug(long double value)
-                -> mass<double, std::micro>
-            {
-                return mass<double, std::micro>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_mg(unsigned long long value) -> milligrams
-            {
-                return milligrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_mg(long double value)
-                -> mass<double, std::milli>
-            {
-                return mass<double, std::milli>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_cg(unsigned long long value) -> centigrams
-            {
-                return centigrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_cg(long double value)
-                -> mass<double, std::centi>
-            {
-                return mass<double, std::centi>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_dg(unsigned long long value) -> decigrams
-            {
-                return decigrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_dg(long double value) -> mass<double, std::deci>
-            {
-                return mass<double, std::deci>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_g(unsigned long long value) -> grams
-            {
-                return grams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_g(long double value)
-                -> mass<double, std::ratio<1>>
-            {
-                return mass<double, std::ratio<1>>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_dag(unsigned long long value) -> decagrams
-            {
-                return decagrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_dag(long double value)
-                -> mass<double, std::deca>
-            {
-                return mass<double, std::deca>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_hg(unsigned long long value) -> hectograms
-            {
-                return hectograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_hg(long double value)
-                -> mass<double, std::hecto>
-            {
-                return mass<double, std::hecto>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_kg(unsigned long long value) -> kilograms
-            {
-                return kilograms(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_kg(long double value) -> mass<double, std::kilo>
-            {
-                return mass<double, std::kilo>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Mg(unsigned long long value) -> megagrams
-            {
-                return megagrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Mg(long double value) -> mass<double, std::mega>
-            {
-                return mass<double, std::mega>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Gg(unsigned long long value) -> gigagrams
-            {
-                return gigagrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Gg(long double value) -> mass<double, std::giga>
-            {
-                return mass<double, std::giga>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Tg(unsigned long long value) -> teragrams
-            {
-                return teragrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Tg(long double value) -> mass<double, std::tera>
-            {
-                return mass<double, std::tera>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Pg(unsigned long long value) -> petagrams
-            {
-                return petagrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Pg(long double value) -> mass<double, std::peta>
-            {
-                return mass<double, std::peta>(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Eg(unsigned long long value) -> exagrams
-            {
-                return exagrams(value);
-            }
-
-            [[nodiscard]] constexpr auto operator""_Eg(long double value) -> mass<double, std::exa>
-            {
-                return mass<double, std::exa>(value);
-            }
-
-        } // namespace mass_literals
-
         inline namespace electric_current_literals {
 
             [[nodiscard]] constexpr auto operator""_aA(unsigned long long value) -> attoamperes
